# docs/v2/openapi.yaml
...
openapi: 3.1.1
info:
  title: Glowing Telegram API
  version: "2.0"
components:
  schemas:
...
    TwitchAuthRequest:
      $ref: "./schemas/TwitchAuthRequest.schema"
    AuthorizationUrlResponse:
      $ref: "./schemas/AuthorizationUrlResponse.schema"
...
paths:
...
  /auth/twitch/url:
    post:
      summary: Obtain Twitch authorization URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwitchAuthRequest"
      responses:
        "200":
          description: Authorization URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationUrlResponse"
  /auth/twitch/callback:
    get:
      summary: Handle Twitch callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to client
          headers:
            Location:
              schema:
                type: string
...